src/main/java/com/weatherapp/dto/WindDetails.java
package com.weatherapp.dto;
               
public class WindDetails
{
    private String deg;

    private String speed;
    
    private String gust;

    public String getDeg ()
    {
        return deg;
    }

    public void setDeg (String deg)
    {
        this.deg = deg;
    }

    public String getSpeed ()
    {
        return speed;
    }

    public void setSpeed (String speed)
    {
        this.speed = speed;
    }

	public String getGust() {
		return gust;
	}

	public void setGust(String gust) {
		this.gust = gust;
	}

}